+++ api/fe.js
@@ -1,0 +1,7 @@
+const { generateModelScript } = require('../forward_engineering/generateModelScript');
+const { validate } = require('../forward_engineering/validation/validate');
+
+module.exports = {
+	generateModelScript,
+	validate,
+};
